move domain to lib
